refactored the PUT router to assignment-service updateAssignment function
